prima/route-planning-service/requirements.txt
Flask==1.1.4
gunicorn==19.9.0
six==1.12.0
pyjwt==1.7.1
flask-marshmallow==0.10.1
python-dotenv==0.10.3
python-jose-cryptodome==1.3.2
#pycryptodome==3.9.0
flask-cors==3.0.8
pandas==1.3.5
openpyxl==3.0.9
ortools==9.1.9490
geopy==2.2.0
pytz==2021.3
memory-profiler==0.60.0
markupsafe==2.0.1
psycopg2-binary==2.8.3
grpcio-tools==1.27.2
google-cloud-pubsub==1.5.0
google-cloud-storage==1.28.1
google-cloud-core==1.3.0
google-cloud-tasks==1.5.0
https://storage.googleapis.com/gobolt-toolbox/python-grpc-utils/python-grpc-utils1.2.5.12.tar.gz
https://storage.googleapis.com/gobolt-toolbox/viper/viper1.0.3.4.tar.gz
https://storage.googleapis.com/gobolt-toolbox/route-planner/dist/route-planner-9f79df7.tar.gz
protobuf==3.19.0
scipy==1.7.3
shapely==1.8.4
here-location-services==0.4.0

